+++ src/components/add-folder-btn/index.js
@@ -1,0 +1,10 @@
+function AddFolderBtn() {
+  return (
+    <button className="add-folder-btn">
+      <p>폴더 추가</p>
+      <img src="/add-white.svg" alt="+ 아이콘" />
+    </button>
+  );
+}
+
+export default AddFolderBtn;
