+++ src/components/card-list/style.css
@@ -50,7 +50,7 @@
   letter-spacing: -0.2px;
 }
 
-.selecter-folder-menu {
+.selected-folder-menu {
   display: flex;
   justify-content: center;
   align-items: center;
